+++ project_cylon/PageFactory.py
@@ -49,8 +49,14 @@ class PageFactory:
 
     @classmethod
     def create_page(cls, doc):
-        page = Page(doc['page']['name'], doc['page']['url'])
-        #page.driver = driver
+        page = Page()
+        page.name = doc['page']['name']
+
+        if "url" in doc['page']:
+            page.url = doc['page']['url']
+
+        if "route" in doc['page']:
+            page.route = doc['page']['route']
 
         if "url_paths" in doc['page']:
             for item in doc['page']['url_paths']:
@@ -60,7 +66,6 @@ class PageFactory:
                 if not name in page.url_paths:
                     page.url_paths[name] = path
                 else:
-                    #log.warning("%s\nDuplicate url path name: '%s'" % (page.name, name))
                     log.warning("Duplicate url path name!", {
                         'page': page.name,
                         'path': name
@@ -85,7 +90,6 @@ class PageFactory:
                 if not element.name in page.elements:
                     page.elements[element.name] = element
                 else:
-                    #log.warning("%s\nDuplicate element name: '%s'" % (page.name, element.name))
                     log.warning("Duplicate element name!", {
                         'page': page.name,
                         'element': element.name
@@ -98,7 +102,7 @@ class PageFactory:
         return page
 
     @classmethod
-    def create_pages(cls, path, site):
+    def create_pages(cls, path, domain):
         pages = {}
 
         for filename in glob.glob(path):
@@ -107,14 +111,30 @@ class PageFactory:
 
             for doc in docs:
                 page = cls.create_page(doc)
-                page.switch_site(site)
+
+                if page.url == "!!undefined":
+                    page.url = cls.build_url(domain, page.route)
 
                 if not page.name in pages:
                     pages[page.name] = page
                 else:
-                    #log.failed("Duplicate page name: '%s'" % page.name)
-                    log.warning("Duplicate page name!", {
-                        'page': page.name
-                    })
+                    log.warning("Duplicate page name!", { 'page': page.name })
 
         return pages
+
+    @classmethod
+    def get_domain(cls, filename, site):
+        content = open(filename, "r")
+        doc = yaml.load(content)
+
+        if site in doc['sites']:
+            return doc['sites'][site]
+        else:
+            return ""
+
+    @classmethod
+    def build_url(cls, domain, path):
+        if domain[-1] == '/': domain = domain[:-1]
+        if path[0] == '/': path = path[1:]
+
+        return "%s/%s" % (domain, path)